Fix issue with highlight group creation
Cleared hl groups will now be recreated properly when changing the color
scheme.

When the ColorScheme autocmd was called, the hl groups wouldn't be
recreated because s:CreateHiGroup() only ran hlexists() to check if a hl
group existed. Cleared hl groups still exist even though they don't have
any highlighting properties, so a helper function has been created which
checks if the hl group exists AND if it's cleared or not.