Update autocmds and statusline generation
This is an update which later will allow us to create statuslines for
any of vim's modes (see ':help mode()' for details). This is
accomplished by executing a function (Pl#GetStatusline()) in every
statusline (mode() only returns the correct mode when being executed
from the statusline).

This update may also make it possible to dynamically generate the
correct highlighting groups for each segment divider and thereby resolve
the issue with incorrect highlighting for the divider between optional
segments surrounded by two segments with different background color.