Only generate BDF or OTF fonts
FontForge fails to generate valid PCF fonts, so pure bitmap fonts will
need to be saved as BDF fonts. They may optionally be converted to PCF
fonts at a later point using `bdftopcf`, but this is usually not
necessary.

TrueType fonts with or without bitmaps will always be stored as OTF
fonts. This may later be updated to be configurable via a command-line
switch.

This commit also changes how the output filename is generated. Only the
filename of the original font up to the first period is used, and
`-Powerline.[format]` is appended to this filename.

Closes #9.