Make sure that hi_curr is set in GroupHandler
Stored statuslines may skip highlighting for several modes so the
statuslines for these modes should be stored as empty statuslines
instead.

If mode highlighting is missing for a mode (e.g. noncurrent, insert),
hi_curr wasn't set and caused an error in GroupHandler. This commit
makes sure that the hi_curr varable is set and will remain empty if no
highlighting is provided in GroupHandler.