Update font patcher
This commit introduces some changes to how the font metrics are
calculated as well as various fixes for glyph alignment and overlap. The
update uses the Latin Extended-A and Box Drawing glyphs to determine the
largest glyphs, this seems to be a quite reliable way of determining
both the glyph height and the maximum width.

An overlap property has been added to SYM_ATTR, the glyphs will be moved
`font.em / 48` in the direction of the glyph alignment so it overlaps
with the previous/next glyph. This is necessary for the block arrows so
no gaps occur.