Create most of the missing theme functionality
This is unfortunately a quite big and messy commit, but all the
functionality depends on each other. This commit introduces the missing
functions for creating colorschemes, matches, segments and themes.

The theme parser is nearly complete and should return all the data
required for generating mode- and buffer-specific statuslines with
selectable highlighting.