Autoload all the files!
This commit changes the plugin structure significantly. Colorscheme,
match, segment, theme and function definitions are now autoloaded, so
only the necessary files and functions are loaded (and they're loaded
automagically which makes everything a lot easier).

Loading the theme dict is now as simple as referring to e.g.
`g:Powerline#Themes#distinguished#theme`. This will load all the
necessary functions and return the full theme data dict.

By autoloading everything it will probably also be easier to implement
customizable theme variables because it's possible to define these
variables _before_ Powerline is loaded, and allow Powerline to use these
variables directly (this also bypasses the autoloading, which may make
the plugin faster for people who customize it in their vimrc).