Handle dynamic insertion/removal of segments
It's important to not refer to any Pl* themes in the vimrc or during the
load process, because plugins may not be loaded at that point and thus
will be disabled.

To be able to dynamically insert/remove segments, a temporary "action
queue" is used (Pl#Mod#theme), and parsed in Pl#Theme#BufferSegments
before the segments are fetched.