Redraw after updating the statusline
This has previously caused some speed issues, but because it's only
triggered when entering/leaving a buffer (not on CursorHold and similar
autocmds) I hope it won't make the plugin any slower. Without redrawing
the statusline will be blank when opening some file types on some
configs, this commit solves that by forcing a redraw after Powerline is
loaded.