Update color scheme syntax
The previous syntax was verbose and difficult to work with. This commit
tries to improve colorscheme creation by simplifying the syntax with the
cost of being easier to break and harder to debug if anything is wrong
in the color scheme.

The new syntax works as following:

First, you need to allocate some colors for your color scheme:

	call Pl#Hi#Allocate({
		\ 'red' : 124,
		\ 'blue': 31,
		\ })

Cterm colors must always be specified. GUI colors are optional. To add
GUI colors, provide a list as the color parameter with the GUI color as
its second element (note the hex notation):

	call Pl#Hi#Allocate({
		\ 'red' : [124, 0xff0000],
		\ 'blue': [31 , 0x0000ff],
		\ })

Now you can create styles for the different segments (example from the
distinguished color scheme):

	let g:Powerline#Colorschemes#distinguished#colorscheme = Pl#Colorscheme#Init([
		\ Pl#Hi#Segments(['mode_indicator'], {
			\ 'n': ['darkestgreen', 'brightgreen', ['bold']],
			\ 'i': ['darkestcyan', 'white', ['bold']],
			\ 'v': ['darkred', 'brightorange', ['bold']],
			\ 'r': ['white', 'brightred', ['bold']],
			\ 's': ['white', 'gray5', ['bold']],
			\ }),
		\ ])

Pl#Colorscheme#Init() accepts a list of segment highlighting properties
(return values from Pl#Hi#Segments()).

Pl#Hi#Segments([...], { ... }) accepts two parameters: a list of all the
segments that are going to be highlighted and a dict with the
highlighting for different modes. Currently 6 modes are available, with
both a shorthand and full key you can use (choose whatever you prefer):

	n | normal
	N | noncurrent
	i | insert
	v | visual
	r | replace
	s | select

The dict values are a list with the colors and attributes for this
segment/mode combination.

The first list element is always the foreground color. The second
element is an optional background color (it's optional for segment
group children--if it's not provided then the segment group's bg color
will be used). The last element in the list is a list of attributes. The
supported attributes are 'bold', 'italic' and 'underline'.

Note: It's important to include highlighting for all the segments in the
default distribution. Failing to include highlighting for used segments
will trigger error messages when starting vim.

Thanks to @sjl for giving me feedback about the new syntax.