Show shortened path in NERDTree buffer
NERDTree (and maybe other buffers in the future) now only shows the most
important parts of the current path. $HOME is shortened to "~" and if
the path consists of more than `threshold` (currently 4) directories it
will be shortened, i.e.
"/home/user/a/very/long/path/indeed" will be shortened to
"~/…/long/path/indeed".

Closes #96.