Added update to SFNT Table for font renaming
On Mac (at least) the font was being grouped with the original font, and
causing confusion for some programs.  This change updates the sfnt_names
table 'Preferred Family' and 'Compatible Full' fields to match the
new font name.  This will cause it to show up as it's own font in the
list on Mac.